app/api/requirements.txt
annotated-types
asttokens
backcall
beautifulsoup4
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
cryptography
decorator
exceptiongroup
executing
Flask
Flask-Cors
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
importlib_metadata
ipython
itsdangerous
jedi
Jinja2
MarkupSafe
matplotlib-inline
parso
pdfminer.six
pdfx
pexpect
pickleshare
prompt-toolkit
proto-plus
protobuf
ptyprocess
pure-eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyparsing
PyPDF2
requests
rsa
six
sortedcontainers
soupsieve
stack-data
tqdm
traitlets
typing_extensions
uritemplate
urllib3
wcwidth
Werkzeug
zipp